requirements.txt
firebase-admin
fastapi
uvicorn
pydantic
python-dotenv
python-jose[cryptography]
python-multipart
email-validator
httpx
